+++ pkg/system/stat_linux.go
@@ -1,0 +1,13 @@
+package system
+
+import (
+	"syscall"
+)
+
+func GetLastAccess(stat *syscall.Stat_t) syscall.Timespec {
+	return stat.Atim
+}
+
+func GetLastModification(stat *syscall.Stat_t) syscall.Timespec {
+	return stat.Mtim
+}
