+++ pkg/system/stat_unsupported.go
@@ -1,0 +1,13 @@
+// +build !linux
+
+package system
+
+import "syscall"
+
+func GetLastAccess(stat *syscall.Stat_t) syscall.Timespec {
+	return stat.Atimespec
+}
+
+func GetLastModification(stat *syscall.Stat_t) syscall.Timespec {
+	return stat.Mtimespec
+}
